+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+## v0.9.0-B190810 (pre-release)
+
 - Improve feedback of parsing errors. [#185](https://github.com/BernieWhite/PSRule/issues/185)
 - Fix outcome reported when error or exception is raised. [#211](https://github.com/BernieWhite/PSRule/issues/211)